WIP: read in "duplicate" plot data for reconcilization. see notes:
We have new data for constructing all post-estimation plots.
At this point, we have two parallel tracks of data making these plots: old and new.
We want our next step to be reconciling the appearance of these plots w/ the new data.
We can then consider how to spin off/improve their presentation, but we want to match the old way for now.